Move footer into main layout.
