+++ src/__tests__/index.jest.test.tsx
@@ -2,7 +2,6 @@ import { render, screen, waitFor, within } from "@testing-library/react"
 import "../__mocks__/routerMock"
 import Home from "@/app/homepage/main"
 import UserEvent from "@testing-library/user-event"
-import { assertFooter } from "./utils/_footer"
 import { assertSocialFab } from "./utils/_socialFab"
 
 describe("Home", () => {
@@ -10,10 +9,9 @@ describe("Home", () => {
     render(<Home termsRead={false} />)
   }
 
-  it("should have a social fab loaded and footer", async () => {
+  it("should have a social fab loaded", async () => {
     renderComponent()
     await assertSocialFab()
-    assertFooter()
   })
 
   it("should be able to click on the cookie button", async () => {
