--- src/app/history/layout.tsx
@@ -1,19 +1,0 @@
-import Footer from "@/components/Footer"
-
-export const metadata = {
-  title: "Website History",
-  description: "Timeline and journey of the page.",
-}
-
-export default function HistoryLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <>
-      {children}
-      <Footer />
-    </>
-  )
-}
