+++ src/app/layout.jest.test.tsx
@@ -2,6 +2,7 @@ import { render, screen } from "@testing-library/react"
 import "../__mocks__/routerMock"
 import { Body } from "./layout"
 import { assertMenu } from "@/__tests__/utils/_menu"
+import { assertFooter } from "@/__tests__/utils/_footer"
 
 describe("Main Layout", () => {
   const renderComponent = () => {
@@ -15,6 +16,7 @@ describe("Main Layout", () => {
   it("should have a Mega Menu", async () => {
     renderComponent()
     await assertMenu()
+    await assertFooter()
     expect(screen.getByText("Sample")).toBeInTheDocument()
   })
 })
