--- src/app/projects/game-snake/layout.tsx
@@ -1,12 +1,0 @@
-export const metadata = {
-  title: "Snake Game",
-  description: "A nokia like snake game.",
-}
-
-export default function SnakeLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <>{children}</>
-}
