--- src/app/projects/lessons/layout.tsx
@@ -1,12 +1,0 @@
-export const metadata = {
-  title: "Lessons",
-  description: "Lesson learned on CI/CD and web developments.",
-}
-
-export default function LessonLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <>{children}</>
-}
