--- src/app/projects/microfrontend/layout.tsx
@@ -1,12 +1,0 @@
-export const metadata = {
-  title: "Microfrontend",
-  description: "Microfrontend architecture.",
-}
-
-export default function MicrofrontEndLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <>{children}</>
-}
