--- src/app/projects/webrtc/layout.tsx
@@ -1,12 +1,0 @@
-export const metadata = {
-  title: "Video Conferencing",
-  description: "Peer to peer Web RTC video conferencing.",
-}
-
-export default function WebrtcLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <>{children}</>
-}
